Move test files to src folder
Place test files with respective components to make them easier to find.
