+++ src/List.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { render } from '@testing-library/react';
 
-import List from '../src/List';
+import List from './List';
 
 test('List', () => {
   const tasks = [{
