--- test/Item.test.jsx
@@ -1,33 +1,0 @@
-import React from 'react';
-
-import { render, fireEvent } from '@testing-library/react';
-
-import Item from '../src/Item';
-
-test('Item', () => {
-  const task = {
-    id: 1,
-    title: '뭐라도 하기',
-  };
-
-  const handleClick = jest.fn();
-
-  const { container, getByText } = render((
-    <Item
-      task={task}
-      onClickDelete={handleClick}
-    />
-  ));
-
-  expect(container).toContainHTML('<li');
-  expect(container).toHaveTextContent('뭐라도 하기');
-
-  expect(container).toContainHTML('<button');
-  expect(container).toHaveTextContent('완료');
-
-  expect(handleClick).not.toBeCalled();
-
-  fireEvent.click(getByText('완료'));
-
-  expect(handleClick).toBeCalledWith(1);
-});
